Refactor additional Interface -> Interfaces
